// web/src/api/models/exploreIcons200.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { ExploreIcons200Links } from './exploreIcons200Links';
import type { ExploreIcons200Meta } from './exploreIcons200Meta';
import type { ExplorerIconResource } from './explorerIconResource';

export type ExploreIcons200 = {
  data?: ExplorerIconResource[];
  links?: ExploreIcons200Links;
  meta?: ExploreIcons200Meta;
};