// web/src/api/models/exploreIconsPlatform.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type ExploreIconsPlatform = typeof ExploreIconsPlatform[keyof typeof ExploreIconsPlatform];


export const ExploreIconsPlatform = {
  ios: 'ios',
  android: 'android',
} as const;