// web/src/api/models/explorerIconResource.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { ExplorerIconResourcePlatform } from './explorerIconResourcePlatform';

export interface ExplorerIconResource {
  app_id?: number;
  external_id?: string;
  platform?: ExplorerIconResourcePlatform;
  name?: string;
  /** @nullable */
  icon_url?: string | null;
  /** @nullable */
  publisher_name?: string | null;
  /** @nullable */
  category_name?: string | null;
}