// web/src/api/models/storeCategoryResource.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { StoreCategory } from './storeCategory';

export type StoreCategoryResource = StoreCategory & ({
  /** @nullable */
  parent_id?: number | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  external_id?: string | null;
});